[refactor] 쿠키 생성 방식 변경

+++ auth-service/src/main/java/com/example/authserver/domain/TokenType.java
@@ -31,11 +31,11 @@ public enum TokenType {
 
     private final Supplier<Instant> instant;
     @Getter
-    private final long Seconds;
+    private final int Seconds;
     private final Function<JwtUtilImpl, String> secret;
 
 
-    TokenType(Supplier<Instant> instant, long toSeconds, Function<JwtUtilImpl, String> secret) {
+    TokenType(Supplier<Instant> instant, int toSeconds, Function<JwtUtilImpl, String> secret) {
         this.instant = instant;
         this.Seconds = toSeconds;
         this.secret = secret;
